Refactor code to use instance of Serializer. Starting to fix #17
